// types.ts
export type TouristAttractions = {
  id: string;
  name: string;
  description: string;
  location: string;
  latitude: number;
  longitude: number;
  image_url: string;
};

export type Restaurants = {
  id: string;
  name: string;
  description: string;
  location: string;
  phone: string;
  latitude: number;
  longitude: number;
  image_url: string;
};
export type Cafes = {
  id: string;
  name: string;
  description: string;
  location: string;
  phone: string;
  latitude: number;
  longitude: number;
  image_url: string;
};

export type Temples = {
  id: string;
  name: string;
  description: string;
  location: string;
  latitude: number;
  longitude: number;
  image_url: string;
};

export type Festivals = {
  id: string;
  name: string;
  description: string;
  event_date: string;
  image_url: string;
};